requirements.txt
openai
gradio
python-dotenv